data/synthetic/requirements.txt
faker>=18.0.0
psycopg2-binary>=2.9.0
numpy>=1.24.0
python-dotenv>=0.21.0
